config: Make leaders_prefernece config hashable
We declare a hash set of values that are restricted in the enterprise wrapper.

This is not excellent, especially since we don't actually use the hash set in
this particular case. It might be better to use a non-associative container
and std::find_if for this purpose - the list of restricted values will tend to
be quite short, and checks will be uncommon and not perf sensitive.