It's a teeny start
Basically we're going to have

- a frontend that replicates a controller command to kick off truncation on
  some number of partitions.
- command batch gets dispatched to some table thing similar to data migrations
- then those notify a migrations backend I guess to go and do the thing.
- probably the table writes our intent to truncate into the kvstore
- then we make some decision about whether we're doing local or TS

For TS we should have something very similar to data migrations backend, it will
just be its own thing. might be kind of redundant, but it is a distinct service

For local we'll probably have a structurally similar flow, but instead of
unmount we shutdown the partition. Still no real idea of where we futz with the
data on disk.

Q: what's the limit for size of controller commands?